reorganizing Forms again, I like this better
