--- src/app/components/Navbar/index.ts
@@ -1,2 +1,0 @@
-export { default as UnauthenticatedNavbar } from './UnauthenticatedNavbar';
-export { default as AuthenticatedNavbar } from './AuthenticatedNavbar';
