+++ src/app/components/common/Navbars/Logo/index.ts
@@ -1,0 +1,1 @@
+export { default as Logo } from './Logo';
