+++ src/app/components/common/Navbars/NavbarDrawer/NavbarDrawerSections/DrawerButtonLink/DrawerButtonLink.tsx
@@ -1,3 +1,4 @@
+'use client';
 import { useCallback } from 'react';
 import Button from '@mui/material/Button';
 import Typography from '@mui/material/Typography';
