+++ src/app/components/common/Navbars/NavbarDrawer/NavbarDrawerSections/index.ts
@@ -1,0 +1,1 @@
+export { MyAccountSection, NavigationSection } from './NavbarDrawerSections';
