+++ src/app/components/common/Navbars/index.ts
@@ -1,0 +1,2 @@
+export { default as AuthenticatedNavbar } from './AuthenticatedNavbar';
+export { default as UnauthenticatedNavbar } from './UnauthenticatedNavbar';
