+++ src/app/components/login-page/LoginForm/LoginForm.tsx
@@ -1,11 +1,11 @@
 'use client';
 import { useActionState } from 'react';
 import { login } from '@/app/data/actions/login';
-import LoginFormHeader from './components/LoginFormHeader';
-import LoginFormBody from './components/LoginFormBody';
-import LoginFormAlert from './components/LoginFormAlert';
-import LoginFormInputs from './components/LoginFormInputs';
-import LoginFormSubmit from './components/LoginFormSubmit';
+import { LoginFormHeader } from './LoginFormHeader';
+import { LoginFormBody } from './LoginFormBody';
+import { LoginFormAlert } from './LoginFormAlert';
+import { LoginFormInputs } from './LoginFormInputs';
+import { LoginFormSubmit } from './LoginFormSubmit';
 
 const LoginForm = () => {
   const [errorMessage, formAction, isPending] = useActionState(login, undefined);
