+++ src/app/components/login-page/LoginForm/LoginFormAlert/index.ts
@@ -1,0 +1,1 @@
+export { default as LoginFormAlert } from './LoginFormAlert';
