+++ src/app/components/login-page/LoginForm/LoginFormBody/index.ts
@@ -1,0 +1,1 @@
+export { default as LoginFormBody } from './LoginFormBody';
