+++ src/app/components/login-page/LoginForm/LoginFormHeader/index.ts
@@ -1,0 +1,1 @@
+export { default as LoginFormHeader } from './LoginFormHeader';
