+++ src/app/components/login-page/LoginForm/LoginFormInputs/index.ts
@@ -1,0 +1,1 @@
+export { default as LoginFormInputs } from './LoginFormInputs';
