+++ src/app/components/login-page/LoginForm/LoginFormSubmit/index.ts
@@ -1,0 +1,1 @@
+export { default as LoginFormSubmit } from './LoginFormSubmit';
