+++ src/app/components/register-page/RegisterForm/RegisterFormAlert/index.ts
@@ -1,0 +1,1 @@
+export { default as RegisterFormAlert } from './RegisterFormAlert';
