+++ src/app/components/register-page/RegisterForm/RegisterFormBody/index.ts
@@ -1,0 +1,1 @@
+export { default as RegisterFormBody } from './RegisterFormBody';
