+++ src/app/components/register-page/RegisterForm/RegisterFormHeader/index.ts
@@ -1,0 +1,1 @@
+export { default as RegisterFormHeader } from './RegisterFormHeader';
