+++ src/app/components/register-page/RegisterForm/RegisterFormInputs/index.ts
@@ -1,0 +1,1 @@
+export { default as RegisterFormInputs } from './RegisterFormInputs';
