+++ src/app/components/register-page/RegisterForm/RegisterFormSubmit/index.ts
@@ -1,0 +1,1 @@
+export { default as RegisterFormSubmit } from './RegisterFormSubmit';
